[main]:EOS:14666:Modularize L10N JSON File Preboarding (#203)
* translation


* translation for preboarding module


* resolve review comment


* resolve review comment
